Migrate test utilities to own package
