+++ packages/react-utils/src/index.ts
@@ -19,6 +19,5 @@ export * from './components/FhirSelect';
 export * from './hooks/utils';
 export * from './components/PageHeader';
 export * from './components/ButtonLink';
-export * from './helpers/test-utils';
 export * from './components/PrivateRoute';
 export * from './components/fhirDataTypes';
