+++ packages/test-utils/README.md
@@ -1,0 +1,46 @@
+# Package
+
+<!-- package description -->
+
+## Installation
+
+```sh
+yarn add @opensrp/template
+```
+
+<!-- 
+Include further installation instructions, for instance if the package requires for the user
+to manually add css.
+-->
+
+## Usage
+
+<!-- 
+
+### Props/ Configuration
+
+#### linkerField
+
+_Optional_(`string` | `undefined` = `undefined`)
+
+When the table is rendered, you can click anywhere on a row to drill down to the next level of the hierarchy. However, you may want to display some kind of indication that it is possible to drill down on a row of data. The `linkerField` prop allows you to define which field should have this indicator. By default this is set to the `id` field.
+
+-->
+
+### Code examples
+
+<!-- 
+
+A single simple, minimal, working snippet for each use-case that the package supports
+
+```tsx
+import { DrillDownTable, columnsFromObjects } from '@onaio/drill-down-table/';
+
+const props = {
+  columns: columnsFromObjects(data),
+  data
+};
+<DrillDownTable {...props} />;
+```
+
+-->
