+++ packages/test-utils/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@opensrp/test-utils",
+  "version": "0.0.0",
+  "description": "",
+  "main": "dist/index.js",
+  "types": "dist/types",
+  "files": [
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": "https://github.com/opensrp/web",
+  "scripts": {
+    "test": "run -T test $INIT_CWD --verbose",
+    "tsc": "run -T tsc",
+    "lint": "run -T eslint ./**/*.{js,jsx,ts,tsx}",
+    "copy": "run -T copyfiles -u 1 \"./src/**/*.{css,html}\" \"./dist/\"",
+    "build": "run tsc && run transpile && run copy",
+    "transpile": "run -T babel src -d dist --root-mode upward --extensions .ts,.tsx --ignore '**/*.test.ts,**/*.test.tsx,**/tests,**/__tests__'"
+  },
+  "jest": {
+    "automock": false,
+    "setupFiles": [
+      "../../setupTests"
+    ]
+  },
+  "bugs": {
+    "url": "https://github.com/opensrp/web/issues"
+  },
+  "dependencies": {
+    "@onaio/connected-reducer-registry": "^0.0.3",
+    "@opensrp/pkg-config": "^0.0.9",
+    "@opensrp/rbac": "workspace:^"
+  },
+  "peerDependencies": {
+    "@opensrp/store": "^0.0.10",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^17.0.0 ",
+    "react-redux": "^7.2.1",
+    "react-router": "^5.2.1",
+    "react-router-dom": "^5.2.1"
+  },
+  "author": "OpenSRP Engineering",
+  "license": "Apache-2.0"
+}
